Added up to scene #4, fix stairs bug, fix smallheart's drop animation

+++ 05-ScenceManager/BackGround.h
@@ -17,7 +17,7 @@ class Background
 public:
 	vector<LPANIMATION> tiles;
 	string path;
-	int screen_limit;
+	float screen_limit;
 	void SetPath(string s)
 	{
 		path = s;
